refactor: create login component
